Removed live JS update as it doesnt work in Firefox & others.
The links can be added dynamically using Javascript, but *not removed* (at least in
Firefox, see this bug: https://bugzilla.mozilla.org/show_bug.cgi?id=380639), so
the list keeps growing. That's why this is removed. If we find a workaround,
or this 2007-bug is fixed (one can hope...), we can revert this commit.